// src/generated/types.ts
export interface paths {
// ...
    "/plans": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Plans
         * @description Fetch a list of plans.
         */
        get: {
            parameters: {
                query?: {
                    type?: "standard" | "metered";
                    currency?: string;
                    active?: boolean;
                    limit?: number;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved the list of plans. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @default true */
                            success: boolean;
                            data?: {
                                plans?: components["schemas"]["Plan"][];
                            };
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        put?: never;
        /**
         * Create Plan
         * @description Create a new plan.
         */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            /** @description The details required to create a plan. */
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PostPlanRequest"];
                };
            };
            responses: {
                /** @description Successfully created plan. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["DataResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/plans/{plan_id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Plan
         * @description Fetch plan information.
         */
        get: {
            parameters: {
                query?: {
                    active?: boolean;
                };
                header?: never;
                path: {
                    plan_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved plan details. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @default true */
                            success: boolean;
                            data?: components["schemas"]["Plan"];
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        /**
         * Update Plan
         * @description Modify plan information.
         */
        put: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    plan_id: string;
                };
                cookie?: never;
            };
            /** @description The details required to update a plan. */
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PutPlanRequest"];
                };
            };
            responses: {
                /** @description Successfully updated plan. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["BaseResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        post?: never;
        /**
         * Remove Plan
         * @description Permanently delete a plan.
         */
        delete: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    plan_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully removed plan. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["BaseResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/plans/{plan_id}/options": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Add Price Option
         * @description Add a new price option to a plan.
         */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    plan_id: string;
                };
                cookie?: never;
            };
            /** @description The details required to create a price option. */
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PostPlanOptionRequest"];
                };
            };
            responses: {
                /** @description Successfully created plan options. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["BaseResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}